third_party/blink/web_tests/accessibility/selection-change-notification-on-selection-removed.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="textbox" role="textbox" contenteditable>
  <p>
      One paragraph in an ARIA textbox.
    </p>
</div>

<script>
  setup(() => {
    let textbox = document.getElementById('textbox');
    textbox.focus();
    let selectionRange = document.createRange();
    selectionRange.selectNodeContents(textbox);
    getSelection().removeAllRanges();
    getSelection().addRange(selectionRange);
  });

  async_test((t) => {
    let axRoot = accessibilityController.rootElement;
    let selectionNotificationsCounter = 0;

    axRoot.addNotificationListener(t.step_func((notification) => {
      if (notification == 'SelectedTextChanged' || notification == 'DocumentSelectionChanged')
        ++selectionNotificationsCounter;

      // A selection change notification should fire four times:
      // 1. When the textbox receives focus.
      // 2. When the selection in removed before it's set via getSelection().removeAllRanges().
      // 3. When the selection is made.
      // 4. When the selection is removed.
      if (selectionNotificationsCounter == 4) {
        axRoot.removeNotificationListener();
        assert_equals(axRoot.selectionAnchorObject, null, 'selectionAnchorObject');
        assert_equals(axRoot.selectionAnchorOffset, -1, 'selectionAnchorOffset');
        assert_equals(axRoot.selectionFocusObject, null, 'selectionFocusObject');
        assert_equals(axRoot.selectionFocusOffset, -1, 'selectionFocusOffset');
        t.done();
      }
    }));

    getSelection().removeAllRanges();
  }, "Tests that a 'selected text changed' notification fires on the document object when the selection is removed.");
</script>
